# openapi-atspaces.yaml
openapi: 3.0.3
info:
  title: At Spaces API
  version: 1.0.0
  description: |
    API contract for At Spaces (NestJS + Prisma + PostgreSQL + Redis).
    Roles: customer, vendor, admin. Access token in JSON, refresh token in HttpOnly cookie.
servers:
  - url: http://localhost:3000/api
  - url: https://<domain>/api
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PagedResponse:
      type: object
      properties:
        items:
          type: array
          items: { type: object }
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        hasNext: { type: boolean }
...
    BranchListItem:
      type: object
      properties:
        id: { type: integer, example: 10 }
        name: { type: string, example: "Branch A" }
        city: { type: string, example: "Amman" }
        address: { type: string, example: "..." }
    FacilityPublic:
      type: object
      properties:
        id: { type: integer, example: 1 }
        name: { type: string, example: "WiFi" }
        icon: { type: string, example: "wifi" }
        isAvailable: { type: boolean, example: true }
        description: { type: string, nullable: true }
    VendorServiceInBranch:
      type: object
      properties:
        vendorServiceId: { type: integer, example: 55 }
        serviceId: { type: integer, example: 1 }
        name: { type: string, example: "Hot Desk" }
        pricePerUnit: { type: number, format: float, example: 5.0 }
        priceUnit: { type: string, enum: [hour, day, week, month], example: hour }
        maxCapacity: { type: integer, example: 50 }
        isAvailable: { type: boolean, example: true }
    BranchDetails:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: string, nullable: true }
        city: { type: string }
        address: { type: string }
        latitude: { type: number, format: float, nullable: true }
        longitude: { type: number, format: float, nullable: true }
        facilities:
          type: array
          items: { $ref: '#/components/schemas/FacilityPublic' }
        services:
          type: array
          items: { $ref: '#/components/schemas/VendorServiceInBranch' }
    AvailabilityCheckRequest:
      type: object
      required: [vendorServiceId, startTime, endTime, quantity]
      properties:
        vendorServiceId: { type: integer, example: 55 }
        startTime: { type: string, format: date-time }
        endTime: { type: string, format: date-time }
        quantity: { type: integer, minimum: 1, example: 2 }
...
    CreateBookingRequest:
      allOf:
        - $ref: '#/components/schemas/AvailabilityCheckRequest'
        - type: object
          required: [paymentMethod]
          properties:
            paymentMethod: { type: string, enum: [cash, card, apple_pay] }
    CreateBookingResponse:
      type: object
      properties:
        bookingId: { type: integer, example: 900 }
        bookingNumber: { type: string, example: "BKG-20260305-0001" }
        totalPrice: { type: number, format: float, example: 20.0 }
        status: { type: string, example: pending }
        paymentStatus: { type: string, example: pending }
...
    VendorBranchUpdateRequest:
      type: object
      required: [name, city, address]
      properties:
        name: { type: string }
        description: { type: string, nullable: true }
        city: { type: string }
        address: { type: string }
        latitude: { type: number, format: float, nullable: true }
        longitude: { type: number, format: float, nullable: true }
    UpdatePriceRequest:
      type: object
      required: [pricePerUnit, priceUnit]
      properties:
        pricePerUnit: { type: number, format: float }
        priceUnit: { type: string, enum: [hour, day, week, month] }
    CapacityRequestRequest:
      type: object
      required: [newCapacity, reason]
      properties:
        newCapacity: { type: integer, minimum: 1 }
        reason: { type: string, minLength: 3 }
    AvailabilityUpsertRequest:
      type: object
      required: [vendorServiceId, date, slots]
      properties:
        vendorServiceId: { type: integer }
        date: { type: string, format: date }
        slots:
          type: array
          items:
            type: object
            required: [start, end, availableUnits]
            properties:
              start: { type: string, example: "10:00" }
              end: { type: string, example: "12:00" }
              availableUnits: { type: integer, minimum: 0 }
    UpdateBookingStatusRequest:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [completed, no_show] }
...
paths:
...
  /branches/{id}:
    get:
      tags: [Branches]
      summary: Branch details (facilities + services)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Branch details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BranchDetails' }
...
  /bookings:
    post:
      tags: [Bookings]
      summary: Create booking
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateBookingRequest' }
      responses:
        '201':
          description: Created booking
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreateBookingResponse' }
...
  /vendors/dashboard:
    get:
      tags: [Vendors]
      summary: Vendor dashboard overview
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Dashboard
          content:
            application/json:
              schema:
                type: object
                properties:
                  todayOccupancy: { type: integer }
                  upcomingBookings: { type: integer }
                  branchStatus: { type: string, enum: [calm, moderate, busy] }
  /vendors/branches/me:
    get:
      tags: [Vendors]
      summary: Vendor branches (owned)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Branches
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/BranchListItem' }
  /vendors/branches/{id}:
    put:
      tags: [Vendors]
      summary: Update vendor branch info
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VendorBranchUpdateRequest' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BranchDetails' }
  /vendors/vendor-services:
    get:
      tags: [Vendors]
      summary: List vendor services
      security: [{ bearerAuth: [] }]
      parameters:
        - name: branchId
          in: query
          required: false
          schema: { type: integer }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Paged vendor services
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PagedResponse' }
  /vendors/vendor-services/{id}:
    get:
      tags: [Vendors]
      summary: Vendor service details
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Vendor service
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VendorServiceInBranch' }
  /vendors/vendor-services/{id}/price:
    put:
      tags: [Vendors]
      summary: Update vendor service price
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdatePriceRequest' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VendorServiceInBranch' }
  /vendors/vendor-services/{id}/capacity-request:
    post:
      tags: [Vendors]
      summary: Create capacity change request (approval)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CapacityRequestRequest' }
      responses:
        '202':
          description: Request created
          content:
            application/json:
              schema:
                type: object
                properties:
                  requestId: { type: integer }
                  status: { type: string, example: pending }
  /vendors/availability:
    put:
      tags: [Vendors]
      summary: Upsert availability slots
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AvailabilityUpsertRequest' }
      responses:
        '200':
          description: OK
  /vendors/bookings:
    get:
      tags: [Vendors]
      summary: Vendor bookings list
      security: [{ bearerAuth: [] }]
      parameters:
        - name: date
          in: query
          required: false
          schema: { type: string, format: date }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: Paged bookings
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PagedResponse' }
  /vendors/bookings/{id}/status:
    patch:
      tags: [Vendors]
      summary: Update booking status (vendor only)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateBookingStatusRequest' }
      responses:
        '200':
          description: OK
...